Some systems define ERR as something
Illumos does in: /usr/include/sys/regset.h when
__EXTENSIONS__ is defined, which we need for other
things. Arguably the system is allowed to pollute as much
as it likes with EXTENSIONS defined so lets catch it
before in the tommath header because tommath seems to use it
for flow control.

Based on Gordon Ross <gwr@nexenta.com> patch posted to samba-technical